Include review feedback

+++ community/kernel/src/main/java/org/neo4j/kernel/builtinprocs/SortedLabels.java
@@ -65,6 +65,6 @@ public class SortedLabels
 
     public Integer label( int offset )
     {
-        return (int) labels[offset];
+        return labels[offset];
     }
 }
